Major refactoring - will reconnect everything
